refactor: Category Entity 관련 수정(Id 속성 int로 변경)

+++ api-server/src/main/java/com/ThreeK_Project/api_server/domain/restaurant/dto/RestaurantRequest.java
@@ -20,5 +20,5 @@ public class RestaurantRequest {
     // 가게 위치 id
     private int locationId;
     // 가게 카테고리 id
-    private Long categoryId;
+    private int categoryId;
 }
